requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.21.0
librosa==0.10.1
numpy==1.26.2
twilio==8.10.0
python-dotenv==1.0.0
soundfile==0.12.1
requests==2.31.0
